Derek:
This is a test of sending commit messages to an email "list".
Actually, it's just an email alias at the moment; the mailing list
will follow.  What you're seeing here would normally be the commit
message.

You should receive two copies of this message since I'm checking
in files in two different directories.

Please make a reply to one of these messages.  If everything is set
up correctly, it should go to the gno-devel list.  Thanks.