SHORT_STANDOUT mode
Previously, nroff embeded the formatting escape sequences into the output string and then tried to re-parse them (as VT52) later for calculating word sizes to justify the text.
This, of course has a handful of problems:

VT100, xterm, etc escape codes aren't counted correctly
the buffer can overflow and cause memory trashing
embedded delays in the termcap are printed as raw characters

This fixes it to use SHORT_STANDOUT mode which uses 8-bit characters to indicate font changes (and tputs to print them).